requirements.txt
Flask
Pillow
colorthief
